nutrition_assistant/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
pydantic==2.12.5
gspread==6.2.1
google-auth==2.49.1
requests==2.32.5
